Fix for when trying to load a booking with an outstanding amount to be paid, that is not an int amount.

+++ lib/data/models/pnr.dart
@@ -267,7 +267,7 @@ class PnrModel {
     if (basket.outstanding.amount == '0') {
       validatePayment = true;
     } else {
-      if( int.parse(basket.outstanding.amount) <= 0 ){
+      if( double.parse(basket.outstanding.amount) <= 0 ){
         print('validatePayment less than 0;');
         validatePayment = true;
       } else {
